rename reportService to ModerationService
