+++ src/bootstrap/container.ts
@@ -9,7 +9,7 @@ import { GuildService } from '../services/guild.service';
 import { HandlerService } from '../services/handler.service';
 import { JobService } from '../services/job.service';
 import { StoreService } from '../services/store.service';
-import { ReportService } from '../services/report.service';
+import { ModService } from '../services/moderation.service';
 import { StorageService } from '../services/storage.service';
 
 export class Container {
@@ -25,6 +25,6 @@ export class Container {
     this._bottle.service('jobService', JobService);
     this._bottle.service('storeService', StoreService);
     this._bottle.service('storageService', StorageService);
-    this._bottle.service('reportService', ReportService, 'storageService');
+    this._bottle.service('modService', ModService, 'storageService');
   }
 }
